Old 2016 interfaces merged with the new 2017 WPILib

+++ src/edu/wpi/first/wpilibj/hal/SerialPortJNI.java
@@ -1,5 +1,5 @@
 /*----------------------------------------------------------------------------*/
-/* Copyright (c) FIRST 2016. All Rights Reserved.                             */
+/* Copyright (c) FIRST 2016-2017. All Rights Reserved.                        */
 /* Open Source Software - may be modified and shared by FRC teams. The code   */
 /* must be accompanied by the FIRST BSD license file in the root directory of */
 /* the project.                                                               */
@@ -24,7 +24,7 @@ public class SerialPortJNI extends JNIWrapper {
 
   public static native void serialSetFlowControl(byte port, byte flow);
 
-  public static native void serialSetTimeout(byte port, float timeout);
+  public static native void serialSetTimeout(byte port, double timeout);
 
   public static native void serialEnableTermination(byte port, char terminator);
